+++ include/class/ASocket.hpp
@@ -1,70 +1,55 @@
 #ifndef TCP_HPP
-#define TCP_HPP
+# define TCP_HPP
 
-#include <fcntl.h>
-#include <exception>
-#include <netinet/in.h>
-#include <stdio.h>
-#include <sys/event.h>
-#include <sys/socket.h>
-#include <sys/types.h>
-#include <fstream>
-#include <unistd.h>
-#include <vector>
-#include <map>
-#include <string>
-#include <cstring>
-#include <iostream>
-#include "error.hpp"
-#include "log.hpp"
-#include "structure.hpp"
+# include <sys/socket.h>
+# include <netinet/in.h>
+# include <unistd.h>
 
-#define LOOP 1
+# include <cstring>
+# include <stdio.h>
+
+# include "log.hpp"
+# include "structure.hpp"
+# include "error.hpp"
+
+# define LOOP 1
 
 /*
-        [TCP functions using UNIX]
-        socketpair
-        htons
-        htonl
-        nthos
-        ntohl
+	[TCP functions using UNIX]
+	socketpair
+	htons
+	htonl
+	nthos
+	ntohl
 
-        select / poll / epoll / kqueue
-        socket
-        accept
-        listen
+	select / poll / epoll / kqueue
+	socket
+	accept
+	listen
+	
+	send
+	recv
+	
+	bind
+	connect
+	
+	getaddrinfo
+	freeaddrinfo
 
-        send
-        recv
-
-        bind
-        connect
-
-        getaddrinfo
-        freeaddrinfo
-
-        setsockpot
-        getsockname
-        getprotobyname
+	setsockpot
+	getsockname
+	getprotobyname
 */
 
 class ASocket {
+	public:
+		struct sockaddr_in	addr;
+		socklen_t			addrSize;
+		socket_t			sock;
 
-
- public:
-  struct sockaddr_in addr;
-  socklen_t addrSize;
-  socket_t server_socket;
-  socket_t client_socket;
-//   std::vector<Server> sever_list;
-
-  ASocket(void);
-  ASocket(int);
-  virtual ~ASocket(void) = 0;
-  void setAddr(void);
-
-  void setNonBlocking(void);
+		ASocket( void );
+		ASocket( socket_t );
+		virtual ~ASocket( void ) = 0;
 };
-
 
 #endif
