+++ include/class/Server.hpp
@@ -1,31 +1,23 @@
 #ifndef SERVER_HPP
-#define SERVER_HPP
+# define SERVER_HPP
 
-#include "ASocket.hpp"
-#include "Client.hpp"
+# include "ASocket.hpp"
 
-#define MAX_EVENTS 10
+class Server: ASocket {
+	public:
+		vec_config_t	cnf;
+	
+		Server( void );
+		~Server( void );
 
-class Server : ASocket {
- public:
-  Server(void);
-  Server(char *);
-  ~Server(void);
+		void	listening( void );
 
+		const vec_config_t&	config( void ) const { return cnf; }
 
-  void change_events( uintptr_t, int16_t ,
-        uint16_t , uint32_t , intptr_t , void *); 
-  void connect_sever();
-
-  void disconnect_client(int  );
-  struct kevent&  getEventList(int );
-private:
-  int kq;
-  std::map<int, std::string> clients;
-  std::vector<struct kevent> change_list;
-  std::vector<struct kevent> server_list;
-  struct kevent* cur_event;
-  // struct kevent event_list[MAX_EVENTS];
+	private:
+		Server( socket_t );
 };
+
+# include "Client.hpp"
 
 #endif
